requirements.txt
PySide6==6.3.2
Pillow==9.0.1

